Ticker functions in utlils.js
